refactor: move `TransactionListItemProps` to models.ts
