+++ src/app/components/TransactionListItem/TransactionListItem.tsx
@@ -4,17 +4,7 @@ import { Icon } from "app/components/Icon";
 import { Label } from "app/components/Label";
 import React from "react";
 
-export type TransactionListItemProps = {
-	date: string;
-	avatarId: string;
-	type: string;
-	address?: string;
-	walletName?: string;
-	amount: string;
-	fiat: string;
-	variant?: "default" | "compact";
-	onClick?: any;
-};
+import { TransactionListItemProps } from "./models";
 
 export const TransactionListItem = ({
 	date,
@@ -50,7 +40,7 @@ export const TransactionListItem = ({
 		return (
 			<tr
 				onClick={onTxClick}
-				className="border-b cursor-pointer border-theme-neutral-200"
+				className="border-b border-theme-neutral-200 cursor-pointer"
 				data-testid="transaction__row"
 			>
 				<td className="w-20 py-4 mt-1">
@@ -74,7 +64,7 @@ export const TransactionListItem = ({
 	return (
 		<tr
 			onClick={onTxClick}
-			className="border-b cursor-pointer border-theme-neutral-200"
+			className="border-b border-theme-neutral-200 cursor-pointer"
 			data-testid="transaction__row"
 		>
 			<td className="w-48 py-1 text-sm text-theme-neutral-600"> {date} </td>
